# src/go/api-gw/docs/swagger.yaml
basePath: /
definitions:
  models.DtoContextListResponse:
    properties:
      data:
        description: Список контекстов
        items:
          $ref: '#/definitions/models.ModelsContext'
        type: array
      meta:
        description: Метаданные пагинации
        properties:
          limit:
            description: limit
            type: integer
          page:
            description: page
            type: integer
          pages:
            description: pages
            type: integer
          total:
            description: total
            type: integer
        type: object
    type: object
  models.DtoCountResponse:
    properties:
      count:
        description: count
        type: integer
    type: object
...
  models.DtoCreateContextRequest:
    properties:
      description:
        description: Описание контекста
        type: string
      id:
        description: Уникальный идентификатор контекста (первичный ключ в БД)
        type: string
      logs_limits:
        description: Лимиты для логов
        properties:
          disk_storage:
            description: disk storage
            type: integer
          retention_duration:
            description: retention duration
            type: integer
        type: object
      metric_limits:
        description: Лимиты для метрик
        properties:
          disk_storage:
            description: disk storage
            type: integer
          retention_duration:
            description: retention duration
            type: integer
        type: object
      name:
        description: Уникальное имя контекста
        type: string
      port:
        description: Уникальный порт для получения метрик
        type: integer
      transport:
        description: Протокол транспортного уровня для получения метрик
        type: string
    type: object
  models.DtoErrorResponse:
    properties:
      error:
        description: error
        type: string
    type: object
...
  models.DtoSuccessResponse:
    properties:
      message:
        description: message
        type: string
    type: object
  models.DtoUpdateContextRequest:
    properties:
      description:
        description: Новое описание контекста
        type: string
      name:
        description: Новое имя контекста
        type: string
    type: object
...
  models.DtoVectroFreePortsResponse:
    properties:
      ports:
        description: ports
        items:
          type: integer
        type: array
    type: object
...
  models.ModelsContext:
    properties:
      admins_count:
        description: admins count
        type: integer
      created_at:
        description: created at
        type: string
      description:
        description: description
        type: string
      id:
        description: id
        type: string
      log_input_port:
        description: Порт для получения метрик в Vector
        type: integer
      log_input_transport:
        description: Протокол транспортного уровня для получения метрик в Vector
        type: string
      logs_limits:
        allOf:
        - $ref: '#/definitions/models.ModelsDataLimits'
        description: logs limits
      metric_limits:
        allOf:
        - $ref: '#/definitions/models.ModelsDataLimits'
        description: metric limits
      metrics_input_port:
        description: metrics input port
        type: string
      name:
        description: name
        type: string
      operators_count:
        description: operators count
        type: integer
    type: object
  models.ModelsDataLimits:
    properties:
      disk_storage:
        description: disk storage
        type: integer
      retention_duration:
        description: retention duration
        type: integer
    type: object
...
host: localhost:8001
info:
  contact: {}
  title: API Gw
  version: "1.0"
paths:
...
  /v1/contexts:
    get:
      consumes:
      - application/json
      description: Получение списка всех контекстов системы с возможностью пагинации
        и фильтрации
      parameters:
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 10
        description: Количество элементов на странице
        in: query
        name: limit
        type: integer
      - description: Фильтр по названию/id контекста
        in: query
        name: search
        type: string
      - default: context_id
        description: Поле для сортировки default(context_id)
        enum:
        - context_id
        - created_at
        in: query
        name: order_by
        type: string
      - default: desc
        description: Направление сортировки default(desc)
        enum:
        - asc
        - desc
        in: query
        name: order_dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DtoContextListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение списка всех контекстов в системе
      tags:
      - contexts
    post:
      consumes:
      - text/plain
      description: Создание нового контекста с указанным названием
      parameters:
      - description: Название контекста
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.DtoCreateContextRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ModelsContext'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
      security:
      - BearerAuth: []
      summary: Добавление нового контекста в систему
      tags:
      - contexts
  /v1/contexts/{context_id}:
    delete:
      consumes:
      - application/json
      description: Удаление контекста по его ID
      parameters:
      - description: ID контекста
        in: path
        name: context_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DtoSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление контекста
      tags:
      - contexts
    get:
      consumes:
      - application/json
      description: Получение информации о контексте по его ID
      parameters:
      - description: ID контекста
        in: path
        name: context_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ModelsContext'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение контекста по его идентификатору
      tags:
      - contexts
    put:
      consumes:
      - application/json
      description: Обновление названия контекста и/или описания
      parameters:
      - description: ID контекста
        in: path
        name: context_id
        required: true
        type: string
      - description: Информация о контексте
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.DtoUpdateContextRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DtoCountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновление контекста по его идентификатору
      tags:
      - contexts
  /v1/contexts/count:
    get:
      consumes:
      - application/json
      description: Получение общего количества контекстов в системе
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DtoCountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение числа всех контекстов системы
      tags:
      - contexts
  /v1/contexts/free-ports:
    get:
      description: Выводит массив портов доступных для получения логов от МЭ
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DtoVectroFreePortsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.DtoErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение списка свободных портов
      tags:
      - contexts
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"