# src/go/api-gw/pkg/models/dto_vectro_free_ports_response.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// DtoVectroFreePortsResponse dto vectro free ports response
//
// swagger:model dto.VectroFreePortsResponse
type DtoVectroFreePortsResponse struct {

	// ports
	Ports []int64 `json:"ports"`
}

// Validate validates this dto vectro free ports response
func (m *DtoVectroFreePortsResponse) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this dto vectro free ports response based on context it is used
func (m *DtoVectroFreePortsResponse) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *DtoVectroFreePortsResponse) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *DtoVectroFreePortsResponse) UnmarshalBinary(b []byte) error {
	var res DtoVectroFreePortsResponse
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}